Modified fields in dto.HelathProfessionalDto.java.
Both types of users (health professionals and patients) have several
common fields for their user accounts. These common fields are out of
date in HealthProfessionalDto.java but they have been updated in
RegisterDto.java

How can this be reused? This is a TO-DO. Build is broken until the TODO
is fixed.